+++ src/bidiMapper/modules/emulation/EmulationProcessor.ts
@@ -102,9 +102,16 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) => await context.setGeolocationOverride(geolocation),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+
+        await context.setGeolocationOverride(config.geolocation ?? null);
+      }),
     );
     return {};
   }
@@ -138,9 +145,23 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) => await context.setLocaleOverride(locale),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+
+        await Promise.all([
+          context.setLocaleOverride(config.locale ?? null),
+          // Set `AcceptLanguage` to locale.
+          context.setUserAgentAndAcceptLanguage(
+            config.userAgent,
+            config.locale,
+          ),
+        ]);
+      }),
     );
     return {};
   }
@@ -171,9 +192,16 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) => await context.setScriptingEnabled(scriptingEnabled),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+
+        await context.setScriptingEnabled(config.scriptingEnabled ?? null);
+      }),
     );
     return {};
   }
@@ -201,10 +229,61 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) =>
-          await context.setScreenOrientationOverride(params.screenOrientation),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+
+        await context.setViewport(
+          config.viewport ?? null,
+          config.devicePixelRatio ?? null,
+          config.screenOrientation ?? null,
+        );
+      }),
+    );
+    return {};
+  }
+
+  async setScreenSettingsOverride(
+    params: Emulation.SetScreenSettingsOverrideParameters,
+  ): Promise<EmptyResult> {
+    const browsingContexts = await this.#getRelatedTopLevelBrowsingContexts(
+      params.contexts,
+      params.userContexts,
+    );
+
+    for (const browsingContextId of params.contexts ?? []) {
+      this.#contextConfigStorage.updateBrowsingContextConfig(
+        browsingContextId,
+        {
+          screenArea: params.screenArea,
+        },
+      );
+    }
+    for (const userContextId of params.userContexts ?? []) {
+      this.#contextConfigStorage.updateUserContextConfig(userContextId, {
+        screenArea: params.screenArea,
+      });
+    }
+
+    await Promise.all(
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+
+        await context.setViewport(
+          config.viewport ?? null,
+          config.devicePixelRatio ?? null,
+          config.screenOrientation ?? null,
+        );
+      }),
     );
     return {};
   }
@@ -307,9 +386,60 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) => await context.setTimezoneOverride(timezone),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+        await context.setTimezoneOverride(config.timezone ?? null);
+      }),
+    );
+    return {};
+  }
+
+  async setTouchOverride(
+    params: Emulation.SetTouchOverrideParameters,
+  ): Promise<EmptyResult> {
+    const maxTouchPoints = params.maxTouchPoints;
+
+    const browsingContexts = await this.#getRelatedTopLevelBrowsingContexts(
+      params.contexts,
+      params.userContexts,
+      true,
+    );
+
+    for (const browsingContextId of params.contexts ?? []) {
+      this.#contextConfigStorage.updateBrowsingContextConfig(
+        browsingContextId,
+        {
+          maxTouchPoints,
+        },
+      );
+    }
+    for (const userContextId of params.userContexts ?? []) {
+      this.#contextConfigStorage.updateUserContextConfig(userContextId, {
+        maxTouchPoints,
+      });
+    }
+
+    if (params.contexts === undefined && params.userContexts === undefined) {
+      this.#contextConfigStorage.updateGlobalConfig({
+        maxTouchPoints,
+      });
+    }
+
+    await Promise.all(
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+        await context.setTouchOverride(config.maxTouchPoints ?? null);
+      }),
     );
     return {};
   }
@@ -350,9 +480,18 @@ export class EmulationProcessor {
     }
 
     await Promise.all(
-      browsingContexts.map(
-        async (context) => await context.setUserAgentOverride(params.userAgent),
-      ),
+      browsingContexts.map(async (context) => {
+        // Actual value can be different from the one in params, e.g. in case of already
+        // existing more granular setting.
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
+        await context.setUserAgentAndAcceptLanguage(
+          config.userAgent,
+          config.locale,
+        );
+      }),
     );
     return {};
   }
@@ -399,13 +538,14 @@ export class EmulationProcessor {
       browsingContexts.map(async (context) => {
         // Actual value can be different from the one in params, e.g. in case of already
         // existing more granular setting.
-        const emulatedNetworkConditions =
-          this.#contextConfigStorage.getActiveConfig(
-            context.id,
-            context.userContext,
-          ).emulatedNetworkConditions ?? null;
+        const config = this.#contextConfigStorage.getActiveConfig(
+          context.id,
+          context.userContext,
+        );
 
-        await context.setEmulatedNetworkConditions(emulatedNetworkConditions);
+        await context.setEmulatedNetworkConditions(
+          config.emulatedNetworkConditions ?? null,
+        );
       }),
     );
     return {};